requirements.txt
asgiref==3.2.10
Django==3.1.2
python-dotenv==0.14.0
pytz==2020.1
sqlparse==0.3.1
djangorestframework==3.12.1
djoser==2.0.5
psycopg2-binary==2.8.6
django-storages==1.10.1
Pillow==7.2.0
boto3==1.15.13
